+++ handlers/EditNode.go
@@ -10,7 +10,7 @@ type NodeEditHandler struct {
 }
 
 func (self *NodeEditHandler) Get() {
-	nid, _ := self.GetInt(":nid")
+	nid, _ := self.GetInt64(":nid")
 	nid_handler := models.GetNode(nid)
 	self.Data["inode"] = nid_handler
 	self.Data["icategory"] = models.GetCategory(nid_handler.Pid)
@@ -21,8 +21,8 @@ func (self *NodeEditHandler) Get() {
 }
 
 func (self *NodeEditHandler) Post() {
-	nid, _ := self.GetInt(":nid")
-	cid, _ := self.GetInt("categoryid")
+	nid, _ := self.GetInt64(":nid")
+	cid, _ := self.GetInt64("categoryid")
 	uid, _ := self.GetSession("userid").(int64)
 	nid_title := self.GetString("title")
 	nid_content := self.GetString("content")
