+++ handlers/NodeHandler.go
@@ -11,8 +11,8 @@ type NodeHandler struct {
 }
 
 func (self *NodeHandler) Get() {
-	page, _ := self.GetInt("page")
-	nid, _ := self.GetInt(":nid")
+	page, _ := self.GetInt64("page")
+	nid, _ := self.GetInt64(":nid")
 
 	nid_handler := models.GetNode(nid)
 	nid_handler.Views = nid_handler.Views + 1
