+++ handlers/root/RApi.go
@@ -29,7 +29,7 @@ func (self *RApi) Get() {
 			self.Ctx.WriteString(outtimesz)
 		} else {
 
-			editmid, _ := self.GetInt(":editmid")
+			editmid, _ := self.GetInt64(":editmid")
 			img := models.GetFile(editmid)
 			self.Data["img"] = img
 			self.TplNames = "root/gallery_editurl.html"
@@ -41,7 +41,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":delmid"); mid != 0 {
+			if mid, _ := self.GetInt64(":delmid"); mid != 0 {
 				if e := models.DelFile(mid); e != nil {
 					self.Data["MsgErr"] = "删除图片文件失败！"
 				} else {
@@ -59,7 +59,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":nid"); mid != 0 {
+			if mid, _ := self.GetInt64(":nid"); mid != 0 {
 				if e := models.DelNode(mid); e != nil {
 					self.Data["MsgErr"] = "删除节点失败！"
 				} else {
@@ -77,7 +77,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":nid"); mid != 0 {
+			if mid, _ := self.GetInt64(":nid"); mid != 0 {
 				if e := models.DelNode(mid); e != nil {
 					self.Data["MsgErr"] = "删除节点失败！"
 				} else {
@@ -93,7 +93,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":tid"); mid != 0 {
+			if mid, _ := self.GetInt64(":tid"); mid != 0 {
 				if e := models.DelTopic(mid); e != nil {
 					self.Data["MsgErr"] = "删除内容失败！"
 				} else {
@@ -109,7 +109,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Redirect("/root-login", 302)
 		} else {
-			if mid, _ := self.GetInt(":tid"); mid != 0 {
+			if mid, _ := self.GetInt64(":tid"); mid != 0 {
 				if e := models.DelTopic(mid); e != nil {
 					self.Data["MsgErr"] = "删除内容失败！"
 				} else {
@@ -125,7 +125,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Redirect("/root-login", 302)
 		} else {
-			if mid, _ := self.GetInt(":tid"); mid != 0 {
+			if mid, _ := self.GetInt64(":tid"); mid != 0 {
 				if e := models.DelTopic(mid); e != nil {
 					self.Data["MsgErr"] = "删除内容失败！"
 				} else {
@@ -143,7 +143,7 @@ func (self *RApi) Get() {
 		} else {
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
-			tid, _ := self.GetInt(":tid")
+			tid, _ := self.GetInt64(":tid")
 			tid_handler := models.GetTopic(tid)
 			self.Data["catpage"] = "about"
 			self.Data["asidepage"] = "root_about_topic_edit"
@@ -161,7 +161,7 @@ func (self *RApi) Get() {
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
 
-			tid, _ := self.GetInt(":tid")
+			tid, _ := self.GetInt64(":tid")
 			tid_handler := models.GetTopic(tid)
 			self.Data["catpage"] = "services"
 			self.Data["asidepage"] = "root-services-topic-edit"
@@ -179,7 +179,7 @@ func (self *RApi) Get() {
 			var cid int64 = 6
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
-			tid, _ := self.GetInt(":tid")
+			tid, _ := self.GetInt64(":tid")
 			tid_handler := models.GetTopic(tid)
 			self.Data["asidepage"] = "root-contact-topic-edit"
 			self.Data["topic"] = tid_handler
@@ -198,7 +198,7 @@ func (self *RApi) Get() {
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
 
-			tid, _ := self.GetInt(":tid")
+			tid, _ := self.GetInt64(":tid")
 			tid_handler := models.GetTopic(tid)
 			self.Data["catpage"] = "crafts"
 			self.Data["asidepage"] = "root-crafts-topic-edit"
@@ -213,7 +213,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":tid"); mid != 0 {
+			if mid, _ := self.GetInt64(":tid"); mid != 0 {
 				if e := models.DelTopic(mid); e != nil {
 					self.Data["MsgErr"] = "删除内容失败！"
 				} else {
@@ -229,7 +229,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":tid"); mid != 0 {
+			if mid, _ := self.GetInt64(":tid"); mid != 0 {
 				if e := models.DelTopic(mid); e != nil {
 					self.Data["MsgErr"] = "删除内容失败！"
 				} else {
@@ -245,7 +245,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//var cid int64 = 6
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
@@ -262,7 +262,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//var cid int64 = 6
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
@@ -279,7 +279,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//var cid int64 = 6
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
@@ -295,7 +295,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
 
@@ -310,7 +310,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
 
@@ -329,7 +329,7 @@ func (self *RApi) Get() {
 			self.Data["MsgErr"], _ = self.GetSession("MsgErr").(string)
 			self.DelSession("MsgErr")
 
-			tid, _ := self.GetInt(":tid")
+			tid, _ := self.GetInt64(":tid")
 			tid_handler := models.GetTopic(tid)
 			self.Data["catpage"] = "information"
 			self.Data["asidepage"] = "root-information-topic-edit"
@@ -344,7 +344,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":nid"); mid != 0 {
+			if mid, _ := self.GetInt64(":nid"); mid != 0 {
 				if e := models.DelNode(mid); e != nil {
 					self.Data["MsgErr"] = "删除节点失败！"
 				} else {
@@ -360,7 +360,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":nid"); mid != 0 {
+			if mid, _ := self.GetInt64(":nid"); mid != 0 {
 				if e := models.DelNode(mid); e != nil {
 					self.Data["MsgErr"] = "删除节点失败！"
 				} else {
@@ -376,7 +376,7 @@ func (self *RApi) Get() {
 		if sess_role, _ := self.GetSession("userrole").(int64); sess_role != -1000 {
 			self.Ctx.Redirect(302, "/root-login")
 		} else {
-			if mid, _ := self.GetInt(":nid"); mid != 0 {
+			if mid, _ := self.GetInt64(":nid"); mid != 0 {
 				if e := models.DelNode(mid); e != nil {
 					self.Data["MsgErr"] = "删除节点失败！"
 				} else {
@@ -401,7 +401,7 @@ func (self *RApi) Post() {
 			inputs := self.Input()
 			url := inputs.Get("url")
 			//圖片ID
-			editmid, _ := self.GetInt(":editmid")
+			editmid, _ := self.GetInt64(":editmid")
 			img := models.GetFile(editmid)
 
 			if e := models.SetFile(editmid, img.Pid, img.Ctype, img.Filename, img.Content, img.Hash, img.Location, url, img.Size); e != nil {
@@ -424,7 +424,7 @@ func (self *RApi) Post() {
 			stitle := self.GetString("stitle")
 			content := self.GetString("content")
 			uid, _ := self.GetSession("userid").(int64)
-			tid, _ := self.GetInt(":tid")
+			tid, _ := self.GetInt64(":tid")
 
 			file, handler, e := self.GetFile("image")
 
@@ -505,7 +505,7 @@ func (self *RApi) Post() {
 			var cid int64 = 1
 			title := self.GetString("title")
 			content := self.GetString("content")
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//编辑NODE
 			if title != "" && nid != 0 {
 				if e := models.SetNode(nid, title, content, cid, 1); e != nil {
@@ -536,7 +536,7 @@ func (self *RApi) Post() {
 			var cid int64 = 3
 			title := self.GetString("title")
 			content := self.GetString("content")
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//编辑NODE
 			if title != "" && nid != 0 {
 				if e := models.SetNode(nid, title, content, cid, 1); e != nil {
@@ -569,8 +569,8 @@ func (self *RApi) Post() {
 			stitle := self.GetString("stitle")
 			content := self.GetString("content")
 			uid, _ := self.GetSession("userid").(int64)
-			tid, _ := self.GetInt(":tid")
-			nid, _ := self.GetInt("nodeid")
+			tid, _ := self.GetInt64(":tid")
+			nid, _ := self.GetInt64("nodeid")
 			file, handler, e := self.GetFile("image")
 
 			if handler != nil && e == nil {
@@ -650,7 +650,7 @@ func (self *RApi) Post() {
 			var cid int64 = 4
 			title := self.GetString("title")
 			content := self.GetString("content")
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//编辑NODE
 			if title != "" && nid != 0 {
 				if e := models.SetNode(nid, title, content, cid, 1); e != nil {
@@ -683,8 +683,8 @@ func (self *RApi) Post() {
 			stitle := self.GetString("stitle")
 			content := self.GetString("content")
 			uid, _ := self.GetSession("userid").(int64)
-			tid, _ := self.GetInt(":tid")
-			nid, _ := self.GetInt("nodeid")
+			tid, _ := self.GetInt64(":tid")
+			nid, _ := self.GetInt64("nodeid")
 			file, handler, e := self.GetFile("image")
 
 			if handler != nil && e == nil {
@@ -766,8 +766,8 @@ func (self *RApi) Post() {
 			stitle := self.GetString("stitle")
 			content := self.GetString("content")
 			uid, _ := self.GetSession("userid").(int64)
-			tid, _ := self.GetInt(":tid")
-			nid, _ := self.GetInt("nodeid")
+			tid, _ := self.GetInt64(":tid")
+			nid, _ := self.GetInt64("nodeid")
 			file, handler, e := self.GetFile("image")
 
 			if handler != nil && e == nil {
@@ -847,7 +847,7 @@ func (self *RApi) Post() {
 			var cid int64 = 5
 			title := self.GetString("title")
 			content := self.GetString("content")
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//编辑NODE
 			if title != "" && nid != 0 {
 				if e := models.SetNode(nid, title, content, cid, 1); e != nil {
@@ -878,7 +878,7 @@ func (self *RApi) Post() {
 			var cid int64 = 6
 			title := self.GetString("title")
 			content := self.GetString("content")
-			nid, _ := self.GetInt(":nid")
+			nid, _ := self.GetInt64(":nid")
 			//编辑NODE
 			if title != "" && nid != 0 {
 				if e := models.SetNode(nid, title, content, cid, 1); e != nil {
@@ -910,8 +910,8 @@ func (self *RApi) Post() {
 			ftitle := self.GetString("title")
 			content := self.GetString("content")
 			uid, _ := self.GetSession("userid").(int64)
-			tid, _ := self.GetInt(":tid")
-			nid, _ := self.GetInt("nodeid")
+			tid, _ := self.GetInt64(":tid")
+			nid, _ := self.GetInt64("nodeid")
 
 			if ftitle != "" && nid != 0 && content != "" {
 				//保存编辑
