+++ handlers/root/RServicesHandler.go
@@ -56,7 +56,7 @@ func (self *RServicesHandler) Get() {
 func (self *RServicesHandler) Post() {
 	title := self.GetString("title")
 	content := self.GetString("content")
-	nodeid, _ := self.GetInt("nodeid")
+	nodeid, _ := self.GetInt64("nodeid")
 	var cid int64 = 3
 	uid, _ := self.GetSession("userid").(int64)
 	sess_username, _ := self.GetSession("username").(string)
@@ -78,7 +78,7 @@ func (self *RServicesHandler) Post() {
 		switch {
 		case helper.Rex(self.Ctx.Request.RequestURI, "^/root-services-edit/([0-9]+)$"):
 			//编辑POST状态
-			tid, _ := self.GetInt(":tid")
+			tid, _ := self.GetInt64(":tid")
 			file, handler, e := self.GetFile("image")
 			defer file.Close()
 			if handler == nil {
